Create my_broken_bits.pl
One issue I found is the arguments are not sending in the appropriate array or the IP defined at the beginning of the code. I was not sure how to do that in perl so I added the variables again inside the function to do some testing but was not successful in finding the issue. The final print statement is also not working as expected.